📂 create : accordionToggle.js / count.js 파일 생성
♻️ Refactor : 코드 리팩토링
- product.js 함수 분리 및 경로 설정
