+++ src/pages/product/product.html
@@ -5,11 +5,7 @@
     <link rel='icon' type='image/svg+xml' href='/favicon.svg' />
     <meta name='viewport' content='width=device-width, initial-scale=1.0' />
     <title>마켓칼리-제품 목록</title>
-    <script type='module' src='/main.js'></script>
-    <script type='module' src='/src/layout/header/header.js'></script>
-    <script type='module' src='/src/layout/footer/footer.js'></script>
     <script type='module' src='/src/pages/product/product.js'></script>
-    <script type='module' src='/src/pages/product/accordion.js'></script>
   </head>
   <body>
     <c-header></c-header>
