+++ src/pages/product/product.js
@@ -1,5 +1,10 @@
+import '@/styles/style.scss';
+import '@/layout/header/header.js';
+import '@/layout/footer/footer.js';
+import '@/pages/product/accordionToggle.js';
 import getPbImageURL from '@/api/getPbImageURL';
 import pb from '@/api/pocketbase';
+import { countFilterDatas, countTotalProducts } from '@/pages/product/count.js';
 
 const categories = ['food', 'necessity', 'personalcare', 'animal'];
 const priceSections = {
@@ -28,156 +33,10 @@ let deliveryFilter = [];
 let benefitFilter = [];
 let typeFilter = [];
 
-async function countEachDatas(){
-  const products = await pb.collection('products').getFullList();
-  
-  const counts = {
-    categories: {
-      food: 0,
-      necessity: 0,
-      personalcare: 0,
-      animal: 0
-    },
-    priceSections: {
-      section1: 0,
-      section2: 0,
-      section3: 0,
-      section4: 0
-    },
-    deliveryTypes: {
-      '샛별배송': 0,
-      '판매자배송': 0
-    },
-    benefits: {
-      '할인상품': 0,
-      '한정수량': 0,
-      '무료배송': 0
-    },
-    types: {
-      'Karly Only': 0
-    }
-  };
-  
-  products.forEach(product => {
-    const benefits = Array.isArray(product.benefit) ? product.benefit : [product.benefit];
-    // 카테고리 별 count
-    if (counts.categories.hasOwnProperty(product.category)) {
-      counts.categories[product.category]++;
-    }
-    // 가격 구간 별 count
-    if (product.finalPrice < 10000) {
-      counts.priceSections.section1++;
-    }
-    else if (product.finalPrice >= 10000 && product.finalPrice <= 49900) {
-      counts.priceSections.section2++;
-    }
-    else if (product.finalPrice >= 50000 && product.finalPrice <= 99900) {
-      counts.priceSections.section3++;
-    }
-    else if (product.finalPrice >= 100000) {
-      counts.priceSections.section4++;
-    }
-    // 배송 유형 별 count
-    if (counts.deliveryTypes.hasOwnProperty(product.deliver)) {
-      counts.deliveryTypes[product.deliver]++;
-    }
-    // 혜택 별 count
-    benefits.forEach(benefit => {
-      if (counts.benefits.hasOwnProperty(benefit)) {
-        counts.benefits[benefit]++;
-      }
-    })
-    // 유형 별 count
-    if (counts.types.hasOwnProperty(product.type)) {
-      counts.types[product.type]++;
-    }
-  });
-
-  categories.forEach(category => {
-    document.getElementById(`count--${category}`).textContent = counts.categories[category];
-  })
-  Object.keys(counts.priceSections).forEach(section => {
-    document.getElementById(`count--${section}`).textContent = counts.priceSections[section];
-  })
-  document.getElementById('count--delivery-daybreak').textContent = counts.deliveryTypes['샛별배송'];
-  document.getElementById('count--delivery-seller').textContent = counts.deliveryTypes['판매자배송'];
-  document.getElementById('count--discount').textContent = counts.benefits['할인상품'];
-  document.getElementById('count--limited').textContent = counts.benefits['한정수량'];
-  document.getElementById('count--free-shipping').textContent = counts.benefits['무료배송'];
-  document.getElementById('count--karly-only').textContent = counts.types['Karly Only'];
-}
-
-
-async function fetchProducts() {
-  const resetButton = document.querySelector('.accordion__title--button-reset');
-  const params = { sort: sortBy };
-  let filters = [];
-
-  try {
-    if (categoryFilter.length > 0) {
-      filters.push(`(${categoryFilter.map(value => `category="${value}"`).join(' || ')})`);
-    }
-    if (priceFilter.length != 0) {
-      filters.push(`(${priceFilter})`);
-    }
-    if (deliveryFilter.length > 0) {
-      filters.push(`(${deliveryFilter.map(value => `deliver="${value}"`).join(' || ')})`);
-    }
-    if (benefitFilter.length > 0) {
-      filters.push(`(${benefitFilter.map(value => `benefit~"${value}"`).join(' || ')})`);
-    }
-    if (typeFilter.length > 0) {
-      filters.push(`(${typeFilter.map(value => `type="${value}"`).join(' || ')})`);
-    }
-    
-    if (filters.length == 0) {
-      resetButton.disabled = true;
-    }
-    if (filters.length > 0) {
-      params.filter = filters.join(' && ');
-      resetButton.disabled = false;
-      resetButton.addEventListener('click', resetFilters);
-    }
-
-    const products = await pb.collection('products').getFullList(params);
-    return products;
-  } 
-  catch (error) {
-    console.error('Error fetching products:', error.message, error.response);
-    return [];
-  }
-}
-/* 초기화 버튼 클릭 시 필터 리셋 */
-function resetFilters() {
-  const checkboxes = document.querySelectorAll('.checkbox');
-  const radioButtons = document.querySelectorAll('.radio');
-  const counts = document.querySelectorAll('.accordion__head--content-count');
-
-  checkboxes.forEach(checkbox => {
-    checkbox.checked = false;
-  });
-  radioButtons.forEach(radioButton => {
-    radioButton.checked = false;
-  });
-  counts.forEach(count => {
-    count.textContent = 0;
-    count.style.display = 'none';
-  });
-
-  categoryFilter = [];
-  priceFilter = [];
-  deliveryFilter = [];
-  benefitFilter = [];
-  typeFilter = [];
-
-  renderProducts();
-}
-
 function createProductCard(product) {
   const card = document.createElement('div');
   card.classList.add('card');
 
-  // const discountPrice = Math.floor(product.price * (1 - product.ratio * 0.01) / 100) * 100;
   const isDiscounted = product.ratio > 0;
 
   const template = isDiscounted ? `
@@ -228,10 +87,44 @@ function createProductCard(product) {
   return card;
 }
 
-/* 총 @건 */
-function countTotalProducts(count){
-  const countCard = document.getElementById('count__card');
-  countCard.textContent = count;
+async function fetchProducts() {
+  const resetButton = document.querySelector('.accordion__title--button-reset');
+  const params = { sort: sortBy };
+  let filters = [];
+
+  try {
+    if (categoryFilter.length > 0) {
+      filters.push(`(${categoryFilter.map(value => `category="${value}"`).join(' || ')})`);
+    }
+    if (priceFilter.length != 0) {
+      filters.push(`(${priceFilter})`);
+    }
+    if (deliveryFilter.length > 0) {
+      filters.push(`(${deliveryFilter.map(value => `deliver="${value}"`).join(' || ')})`);
+    }
+    if (benefitFilter.length > 0) {
+      filters.push(`(${benefitFilter.map(value => `benefit~"${value}"`).join(' || ')})`);
+    }
+    if (typeFilter.length > 0) {
+      filters.push(`(${typeFilter.map(value => `type="${value}"`).join(' || ')})`);
+    }
+    
+    if (filters.length == 0) {
+      resetButton.disabled = true;
+    }
+    if (filters.length > 0) {
+      params.filter = filters.join(' && ');
+      resetButton.disabled = false;
+      resetButton.addEventListener('click', resetFilter);
+    }
+
+    const products = await pb.collection('products').getFullList(params);
+    return products;
+  } 
+  catch (error) {
+    console.error('Error fetching products:', error.message, error.response);
+    return [];
+  }
 }
 
 async function renderProducts() {
@@ -245,7 +138,7 @@ async function renderProducts() {
   });
   
   countTotalProducts(products.length);
-  countEachDatas();
+  countFilterDatas();
 }
 
 /* 정렬 버튼 */
@@ -271,6 +164,32 @@ function sortProducts() {
       renderProducts();
     });
   });
+}
+
+/* 초기화 버튼 클릭 시 필터 리셋 */
+function resetFilter() {
+  const checkboxes = document.querySelectorAll('.checkbox');
+  const radioButtons = document.querySelectorAll('.radio');
+  const counts = document.querySelectorAll('.accordion__head--content-count');
+
+  checkboxes.forEach(checkbox => {
+    checkbox.checked = false;
+  });
+  radioButtons.forEach(radioButton => {
+    radioButton.checked = false;
+  });
+  counts.forEach(count => {
+    count.textContent = 0;
+    count.style.display = 'none';
+  });
+
+  categoryFilter = [];
+  priceFilter = [];
+  deliveryFilter = [];
+  benefitFilter = [];
+  typeFilter = [];
+
+  renderProducts();
 }
 
 function category(){
